Arreglado para que de números grandísimos
Lo que hice fué no seguirle los condejos a Val xd

+++ Primer_Laboratorio_De_Discretas/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace Primer_Laboratorio_De_Discretas
 {
@@ -12,38 +13,38 @@ namespace Primer_Laboratorio_De_Discretas
         {
             Console.WriteLine("Hey holmes!, introduce a number and let´s get into it!");
             string input = Console.ReadLine();
-            int a = int.Parse(input);
+            BigInteger a = int.Parse(input);
             /*a is the number that the user introduces, it symbolizes the position in the succession*/
-            int integer = a - 1;
+            BigInteger integer = a - 1;
 
-            List<int> pentagonal = new List<int>();
-            List<int> fibonacci = new List<int>() { 0, 1 };
+            List<BigInteger> pentagonal = new List<BigInteger>();
+            List<BigInteger> fibonacci = new List<BigInteger>() { 0, 1 };
             
-            for (int n =1; n < (a + 1); n++)
+            for (BigInteger n =1; n < (a + 1); n++)
             {
-                int penta = (n * ((3 * n) - 1)) / 2;
+                BigInteger penta = (n * ((3 * n) - 1)) / 2;
                 pentagonal.Add(penta);
             }
 
-            for (int y = 1; y < a - 1; y++) /* a-1 is used because there are already two initial elements in the fibonacci list, and we need the number of elements that are introduced on "a" plus 1 from the for*/
+            for (BigInteger y = 1; y < a - 1; y++) /* a-1 is used because there are already two initial elements in the fibonacci list, and we need the number of elements that are introduced on "a" plus 1 from the for*/
             {
                 var fibo = (fibonacci[fibonacci.Count - 1]) + (fibonacci[fibonacci.Count - 2]);
                 fibonacci.Add(fibo);
             }
 
-            int penta_acum = 0;
-            foreach ( int p in pentagonal)
+            BigInteger penta_acum = 0;
+            foreach ( BigInteger p in pentagonal)
             {
                 penta_acum += p; 
             }
 
-            int fibo_acum = 0;
-            foreach (int f in fibonacci)
+            BigInteger fibo_acum = 0;
+            foreach (BigInteger f in fibonacci)
             {
                 fibo_acum += f; 
             }
 
-            int product = penta_acum * fibo_acum;
+            BigInteger product = penta_acum * fibo_acum;
             Console.WriteLine(product);
         }
     }
